# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Audio Visual Layer API
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sessionToken
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: sessionToken
      description: Same token as bearerAuth but additionally requires ADMIN role
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
    ProjectSummary:
      type: object
      required: [id, userId, title, schemaVersion, visibility, createdAt, updatedAt]
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        schemaVersion:
          type: integer
        visibility:
          type: string
          enum: [PRIVATE, PUBLIC]
        thumbnailUrl:
          type: string
          nullable: true
        shareToken:
          type: string
          nullable: true
          description: Present if the project has been published
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Project:
      allOf:
        - $ref: '#/components/schemas/ProjectSummary'
        - type: object
          required: [documentJson]
          properties:
            documentJson:
              description: The full serialized Project document
              type: object
              additionalProperties: true
    CreateProjectInput:
      type: object
      required: [title, documentJson]
      properties:
        title:
          type: string
        documentJson:
          type: object
          additionalProperties: true
        schemaVersion:
          type: integer
          default: 1
        thumbnailUrl:
          type: string
    UpdateProjectInput:
      type: object
      properties:
        title:
          type: string
        documentJson:
          type: object
          additionalProperties: true
        thumbnailUrl:
          type: string
    ProjectListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProjectSummary'
    ProjectResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Project'
    ShareResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [shareToken]
          properties:
            shareToken:
              type: string

    # Password reset
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      security:
        - bearerAuth: []
      tags: [projects]
      summary: List the authenticated user's projects
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      operationId: createProject
      security:
        - bearerAuth: []
      tags: [projects]
      summary: Save a new project to the cloud
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectInput'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{id}:
    get:
      operationId: getProject
      security:
        - bearerAuth: []
      tags: [projects]
      summary: Get a project by ID (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      operationId: updateProject
      security:
        - bearerAuth: []
      tags: [projects]
      summary: Update project title, document, or thumbnail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectInput'
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: deleteProject
      security:
        - bearerAuth: []
      tags: [projects]
      summary: Delete a project and all its assets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{id}/share:
    post:
      operationId: shareProject
      security:
        - bearerAuth: []
      tags: [projects]
      summary: Publish project — creates or returns the share token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Share token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: unshareProject
      security:
        - bearerAuth: []
      tags: [projects]
      summary: Unpublish project — revokes the share token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Share revoked
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Public share viewer ───────────────────────────────────────────────────
...